tasks/elasticsearch_cleaner/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
boto3==1.4.5
botocore==1.5.94          # via boto3, s3transfer
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
docutils==0.14            # via botocore
futures==3.1.1            # via s3transfer
idna==2.5                 # via requests
jmespath==0.9.3           # via boto3, botocore
ply==3.9                  # via pyhcl
pyhcl==0.3.7
python-dateutil==2.6.1    # via botocore
requests==2.18.3
s3transfer==0.1.10        # via boto3
six==1.10.0               # via python-dateutil
urllib3==1.22             # via requests
